feat(scheduling): 优化数据准备线程并添加学时类型缓存
- 将 CreditHourTypeDAO 和 CreditHourTypeLogic 中的 getCreditHourTypeByUuid 方法实现为通过 Redis 缓存学时类型信息
- 优化 ScheduleLessonsDataPreparationThread 中的数据准备逻辑，包括课程时间表项的准备和教室信息的获取
- 修改 CourseScheduleItemDTO 中的 courseEnuType 字段为 courseType，类型从 CourseEnuType改为 CreditHourTypeDTO
- 调整 CourseTypePriorityDTO 中的 priority 字段类型从 short 改为 Integer

+++ src/main/java/com/frontleaves/scheduling/models/dto/merge/CourseLibraryAndTeacherCourseQualificationListDTO.java
@@ -92,5 +92,5 @@ public class CourseLibraryAndTeacherCourseQualificationListDTO {
     /**
      * 课程优先级
      */
-    private Short priority;
+    private Integer priority;
 }
